Using getFromEnv everywhere
